Calculated empirical density and weight based on variable split.

This function calculate an empirical density of raw data based on variable split from Ceteris Paribus profiles. Then calculated weight for values generated by ingredients::ceteris_paribus().

calculate_weight(profiles, data, variable_split)

Arguments

profiles
data.frame generated by ingredients::ceteris_paribus()

data
data.frame with raw data to model

variable_split
list generated by vivo::calculate_variable_split()

Value

Return an weight based on empirical density.

Examples

library("DALEX", warn.conflicts = FALSE, quietly = TRUE)

#> Warning: pakiet 'DALEX' został zbudowany w wersji R 3.6.2

#> Welcome to DALEX (version: 1.0).
#> Find examples and detailed introduction at: https://pbiecek.github.io/ema/

data(apartments)

library("ingredients", warn.conflicts = FALSE, quietly = TRUE)
split <- vivo::calculate_variable_split(apartments,
                        variables = colnames(apartments),
                        grid_points = 101)

library("randomForest", warn.conflicts = FALSE, quietly = TRUE)

#> randomForest 4.6-14

#> Type rfNews() to see new features/changes/bug fixes.

apartments_rf_model <- randomForest(m2.price ~ construction.year + surface +
                                    floor + no.rooms, data = apartments)

explainer_rf <- explain(apartments_rf_model, data = apartmentsTest[,2:5],
                        y = apartmentsTest$m2.price)

#> Preparation of a new explainer is initiated
#>   -> model label       :  randomForest  (  default  )
#>   -> data              :  9000  rows  4  cols 
#>   -> target variable   :  9000  values 
#>   -> model_info        :  package randomForest , ver. 4.6.14 , task regression (  default  ) 
#>   -> predict function  :  yhat.randomForest  will be used (  default  )
#>   -> predicted values  :  numerical, min =  2105.619 , mean =  3515.128 , max =  5346.334  
#>   -> residual function :  difference between y and yhat (  default  )
#>   -> residuals         :  numerical, min =  -1235.195 , mean =  -3.604187 , max =  2168.808  
#>   A new explainer has been created!  

new_apartment <- data.frame(construction.year = 1998, surface = 88, floor = 2L, no.rooms = 3)

profiles <- ceteris_paribus(explainer_rf, new_apartment)

library("vivo")
calculate_weight(profiles, data = apartments[, 2:5], variable_split = split)

#> $construction.year
#>  [1] 0.004 0.011 0.013 0.005 0.012 0.012 0.011 0.016 0.011 0.015 0.003 0.016
#> [13] 0.008 0.011 0.012 0.013 0.011 0.011 0.012 0.008 0.014 0.011 0.007 0.006
#> [25] 0.012 0.013 0.014 0.013 0.000 0.014 0.017 0.011 0.008 0.010 0.008 0.009
#> [37] 0.007 0.017 0.018 0.011 0.016 0.010 0.009 0.010 0.010 0.009 0.013 0.007
#> [49] 0.011 0.016 0.011 0.010 0.018 0.010 0.009 0.006 0.016 0.003 0.014 0.008
#> [61] 0.005 0.011 0.011 0.006 0.013 0.009 0.017 0.009 0.015 0.008 0.012 0.015
#> [73] 0.012 0.015 0.011 0.016 0.009 0.013 0.015 0.012 0.012 0.007 0.009 0.009
#> [85] 0.010 0.015 0.014 0.004 0.010 0.010 0.012 0.003
#> 
#> $surface
#>   [1] 0.009 0.009 0.007 0.026 0.010 0.005 0.008 0.008 0.006 0.015 0.015 0.004
#>  [13] 0.004 0.017 0.017 0.014 0.014 0.014 0.007 0.012 0.012 0.008 0.008 0.015
#>  [25] 0.009 0.014 0.012 0.012 0.011 0.012 0.008 0.011 0.011 0.021 0.006 0.012
#>  [37] 0.012 0.006 0.013 0.013 0.005 0.007 0.016 0.016 0.007 0.006 0.016 0.008
#>  [49] 0.009 0.017 0.009 0.008 0.016 0.016 0.008 0.020 0.020 0.009 0.012 0.012
#>  [61] 0.004 0.007 0.012 0.012 0.007 0.026 0.026 0.010 0.018 0.018 0.009 0.009
#>  [73] 0.014 0.014 0.009 0.008 0.008 0.010 0.017 0.017 0.007 0.017 0.012 0.009
#>  [85] 0.011 0.016 0.004 0.001 0.015 0.015 0.007 0.005 0.019 0.019 0.004 0.016
#>  [97] 0.016 0.010 0.013 0.013
#> 
#> $floor
#>  [1] 0.090 0.116 0.087 0.086 0.095 0.104 0.103 0.103 0.108 0.108
#> 
#> $no.rooms
#> [1] 0.099 0.202 0.231 0.223 0.198 0.047
#>