Calculated empirical density and weight based on variable split.

This function calculate an empirical density of raw data based on variable split from Ceteris Paribus profiles. Then calculated weight for values generated by ingredients::ceteris_paribus().

calculate_weight(profiles, data, variable_split)

Arguments

profiles
data.frame generated by ingredients::ceteris_paribus()

data
data.frame with raw data to model

variable_split
list generated by ingredients::calculate_variable_split()

Value

Return an weight based on empirical density.

Examples

library("DALEX", warn.conflicts = FALSE, quietly = TRUE)

#> Welcome to DALEX (version: 0.3.0).
#> This is a plain DALEX. Use 'install_dependencies()' to get all required packages.

data(apartments)

library("ingredients", warn.conflicts = FALSE, quietly = TRUE)
split <- calculate_variable_split(apartments, variables = colnames(apartments))

library("randomForest", warn.conflicts = FALSE, quietly = TRUE)

#> randomForest 4.6-14

#> Type rfNews() to see new features/changes/bug fixes.

apartments_rf_model <- randomForest(m2.price ~ construction.year + surface +
                                    floor + no.rooms, data = apartments)

explainer_rf <- explain(apartments_rf_model, data = apartmentsTest[,2:5],
                        y = apartmentsTest$m2.price)

new_apartment <- data.frame(construction.year = 1998, surface = 88, floor = 2L, no.rooms = 3)

profiles <- ceteris_paribus(explainer_rf, new_apartment)

library("vivo")
calculate_weight(profiles, data = apartments[, 2:5], variable_split = split)

#> $construction.year
#>  [1] 0.015 0.015 0.013 0.005 0.012 0.012 0.011 0.016 0.011 0.015 0.019 0.019
#> [13] 0.008 0.011 0.012 0.013 0.011 0.011 0.012 0.008 0.014 0.011 0.013 0.013
#> [25] 0.012 0.013 0.014 0.013 0.014 0.014 0.017 0.011 0.008 0.010 0.008 0.009
#> [37] 0.024 0.024 0.018 0.011 0.016 0.010 0.009 0.010 0.010 0.022 0.022 0.018
#> [49] 0.018 0.016 0.011 0.010 0.018 0.010 0.009 0.006 0.016 0.017 0.017 0.008
#> [61] 0.016 0.016 0.011 0.006 0.013 0.009 0.017 0.009 0.015 0.008 0.012 0.015
#> [73] 0.012 0.015 0.011 0.016 0.009 0.013 0.015 0.012 0.012 0.007 0.009 0.009
#> [85] 0.010 0.015 0.014 0.014 0.014 0.010 0.012 0.003
#> 
#> $surface
#>   [1] 0.013 0.013 0.007 0.013 0.010 0.005 0.014 0.014 0.014 0.006 0.009 0.011
#>  [13] 0.014 0.014 0.007 0.014 0.006 0.008 0.013 0.012 0.012 0.008 0.008 0.015
#>  [25] 0.009 0.014 0.004 0.008 0.011 0.012 0.008 0.011 0.011 0.012 0.006 0.014
#>  [37] 0.006 0.012 0.013 0.013 0.005 0.011 0.012 0.011 0.011 0.006 0.016 0.008
#>  [49] 0.009 0.007 0.017 0.017 0.016 0.016 0.008 0.016 0.018 0.018 0.012 0.012
#>  [61] 0.009 0.007 0.012 0.012 0.021 0.015 0.011 0.010 0.009 0.009 0.009 0.009
#>  [73] 0.014 0.014 0.011 0.008 0.008 0.010 0.012 0.011 0.024 0.024 0.005 0.016
#>  [85] 0.011 0.007 0.010 0.010 0.010 0.010 0.007 0.018 0.018 0.009 0.004 0.009
#>  [97] 0.007 0.010 0.009 0.004
#> 
#> $floor
#>  [1] 0.090 0.116 0.087 0.086 0.095 0.104 0.103 0.103 0.108 0.108
#> 
#> $no.rooms
#> [1] 0.301 0.301 0.231 0.223 0.198 0.047
#>