Local Variable Importance measure based on Ceteris Paribus profiles.

This function calculate local importance measure in eight variants. We obtain eight variants measure through the possible options of three parameters such as absolute_deviation, point and density.

local_variable_importance(profiles, data, absolute_deviation = TRUE,
  point = TRUE, density = TRUE, grid_points = 101)

Arguments

profiles
data.frame generated by ingredients::ceteris_paribus()

data
data.frame with raw data to model

absolute_deviation
logical parameter, if absolute_deviation = TRUE then measure is calculated as absolute deviation, else is calculated as a root from average squares

point
logical parameter, if point = TRUE then measure is calculated as a distance from f(x), else measure is calculated as a distance from average profiles

density
logical parameter, if density = TRUE then measure is weighted based on the density of variable, else is not weighted

grid_points
maximum number of points for profile calculations, the default values is 101, the same as in ingredients::ceteris_paribus, if you use a different on, you should also change here

Value

A data.frame of the class local_variable_importance. It's a data.frame with calculated local variable importance measure.

Examples

library("DALEX")
data(apartments)

library("randomForest")
apartments_rf_model <- randomForest(m2.price ~ construction.year + surface +
                                    floor + no.rooms, data = apartments)

explainer_rf <- explain(apartments_rf_model, data = apartmentsTest[,2:5],
                        y = apartmentsTest$m2.price)

#> Preparation of a new explainer is initiated
#>   -> model label       :  randomForest  (  default  )
#>   -> data              :  9000  rows  4  cols 
#>   -> target variable   :  9000  values 
#>   -> model_info        :  package randomForest , ver. 4.6.14 , task regression (  default  ) 
#>   -> predict function  :  yhat.randomForest  will be used (  default  )
#>   -> predicted values  :  numerical, min =  2103.719 , mean =  3514.285 , max =  5364.833  
#>   -> residual function :  difference between y and yhat (  default  )
#>   -> residuals         :  numerical, min =  -1252.104 , mean =  -2.761149 , max =  2139.647  
#>   A new explainer has been created!  

new_apartment <- data.frame(construction.year = 1998, surface = 88, floor = 2L, no.rooms = 3)

library("ingredients")
profiles <- ceteris_paribus(explainer_rf, new_apartment)

library("vivo")
local_variable_importance(profiles, apartments[,2:5],
                          absolute_deviation = TRUE, point = TRUE, density = TRUE)

#>       variable_name   measure
#> 1 construction.year 192.52106
#> 2           surface 310.88900
#> 3             floor 250.30481
#> 4          no.rooms  69.21883

local_variable_importance(profiles, apartments[,2:5],
                          absolute_deviation = TRUE, point = TRUE, density = FALSE)

#>       variable_name   measure
#> 1 construction.year 192.15905
#> 2           surface 272.56384
#> 3             floor 242.72493
#> 4          no.rooms  93.53277

local_variable_importance(profiles, apartments[,2:5],
                          absolute_deviation = TRUE, point = FALSE, density = TRUE)

#>       variable_name   measure
#> 1 construction.year 142.63351
#> 2           surface 327.17922
#> 3             floor 184.98741
#> 4          no.rooms  71.32424